refactor: share edge chart tooltips
