+++ ui/src/composables/charts/shared/edgeGraph.ts
@@ -1,5 +1,6 @@
 import { getNextColorFor } from '@/lib/utils'
 import type { Point3D, SortOrder } from '@/types'
+import { formatTooltipValue } from './chartConfig'
 
 export type EdgeGraphLink = { source: string; target: string; value: number }
 export type EdgeGraphNode = { name: string; itemStyle?: { color?: string }; total: number }
@@ -50,4 +51,25 @@ export function sortEdgeGraphNodes(nodes: EdgeGraphNode[], order: SortOrder): Ed
     if (a.total !== b.total) return multiplier * (a.total - b.total)
     return a.name.localeCompare(b.name)
   })
+}
+
+export function formatEdgeTooltip(params: any, colorFor: (name: string) => string): string {
+  const dot = (color: string) =>
+    `<span style="display:inline-block;width:10px;height:10px;border-radius:50%;background:${color};margin-right:6px"></span>`
+
+  if (params.dataType === 'edge' || params.data?.source != null) {
+    const source = String(params.data?.source ?? params.name ?? '')
+    const target = String(params.data?.target ?? '')
+    const value = params.data?.value ?? params.value
+    return (
+      `${dot(colorFor(source))}<strong>${source}</strong> → ` +
+      `${dot(colorFor(target))}<strong>${target}</strong><br/>${formatTooltipValue(value)}`
+    )
+  }
+
+  const name = String(params.name ?? params.data?.name ?? '')
+  const value = params.value ?? params.data?.value
+  const valueLine =
+    value === undefined || value === null ? '' : `<br/>${formatTooltipValue(value)}`
+  return `${params.marker || dot(colorFor(name))} <strong>${name}</strong>${valueLine}`
 }
