+++ ui/src/composables/charts/useSankeyChartOptions.ts
@@ -2,14 +2,9 @@ import { computed } from 'vue'
 import type { EChartsOption } from 'echarts'
 import { type BaseChartConfig, getBaseOptions } from './baseChartOptions'
 import { hasXAxis, hasYAxis } from '@/lib/utils'
-import { getChartStyling, getTooltipTheme, formatTooltipValue } from './shared'
+import { getChartStyling, getTooltipTheme } from './shared'
 import { fontSize } from './shared/common'
-import { buildEdgeGraph, sortEdgeGraphNodes } from './shared/edgeGraph'
-
-/** ECharts-style color circle for tooltip rows (matches other chart formatters). */
-function tooltipColorDot(color: string): string {
-  return `<span style="display:inline-block;width:10px;height:10px;border-radius:50%;background:${color};margin-right:6px"></span>`
-}
+import { buildEdgeGraph, formatEdgeTooltip, sortEdgeGraphNodes } from './shared/edgeGraph'
 
 /** ECharts defaults right: '20%' for label room; we keep a tighter inset. */
 const sankeyLayout = {
@@ -69,27 +64,7 @@ export function useSankeyChartOptions(config: BaseChartConfig) {
       tooltip: {
         trigger: 'item',
         ...getTooltipTheme(isDark.value),
-        formatter: (params: any) => {
-          // Link hover: colored source → colored target + weight
-          if (params.dataType === 'edge' || params.data?.source != null) {
-            const src = String(params.data?.source ?? params.name ?? '')
-            const tgt = String(params.data?.target ?? '')
-            const val = params.data?.value ?? params.value
-            return (
-              `${tooltipColorDot(colorFor(src))}<strong>${src}</strong>` +
-              ` → ` +
-              `${tooltipColorDot(colorFor(tgt))}<strong>${tgt}</strong>` +
-              `<br/>${formatTooltipValue(val)}`
-            )
-          }
-          // Node hover: name + total flow through the node
-          const name = params.name ?? params.data?.name ?? ''
-          const val = params.value ?? params.data?.value
-          const valueLine =
-            val === undefined || val === null ? '' : `<br/>${formatTooltipValue(val)}`
-          const marker = params.marker || tooltipColorDot(colorFor(String(name)))
-          return `${marker} <strong>${name}</strong>${valueLine}`
-        },
+        formatter: (params: any) => formatEdgeTooltip(params, colorFor),
       } as EChartsOption['tooltip'],
       series: [
         {
